requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.116.0
h11==0.16.0
httptools==0.6.4
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1

# MongoDB 관련 패키지
motor==3.3.2
pymongo==4.6.1
beanie==1.24.0

# OAuth2 및 JWT 관련 패키지
authlib==1.2.1
itsdangerous==2.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2

# 설정 관리
pydantic-settings==2.1.0

# 인증
PyJWT==2.10.1
email-validator==2.2.0

# Socket.IO
python-socketio==5.13.0

# 암호화
cryptography==45.0.5

# 테스트용 라이브러리
pytest==8.4.1
pytest-asyncio==0.23.8
requests==2.32.3
aiohttp==3.9.5
mongomock-motor==0.0.36